refactor(tangle-dapp): move useInitialNetwork to tangle-shared-ui
