+++ libs/tangle-shared-ui/src/hooks/useInitialNetwork.ts
@@ -1,7 +1,3 @@
-import { DEFAULT_NETWORK } from '@webb-tools/tangle-shared-ui/constants/networks';
-import useLocalStorage, {
-  LocalStorageKey,
-} from '@webb-tools/tangle-shared-ui/hooks/useLocalStorage';
 import {
   Network,
   NETWORK_MAP,
@@ -9,8 +5,9 @@ import {
 } from '@webb-tools/webb-ui-components/constants/networks';
 import { useCallback } from 'react';
 import { z } from 'zod';
-
-import testRpcEndpointConnection from '../components/NetworkSelector/testRpcEndpointConnection';
+import useLocalStorage, { LocalStorageKey } from './useLocalStorage';
+import { DEFAULT_NETWORK } from '../constants/networks';
+import testRpcEndpointConnection from '../utils/testRpcEndpointConnection';
 import createCustomNetwork from '../utils/createCustomNetwork';
 
 const useCachedNetworkId = (): ((
